# specs/001-backend-api/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Todo Task Management API
  description: |
    RESTful API for managing todo tasks with JWT authentication and user isolation.

    ## Authentication
    All endpoints require JWT authentication via Bearer token in the Authorization header.
    The JWT token must be issued by Better Auth and contain user_id and email claims.

    ## Security
    - User isolation enforced: users can only access their own tasks
    - JWT tokens expire after 7 days
    - Unauthorized access returns 401, non-existent resources return 404
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com

servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.example.com
    description: Production server

security:
  - BearerAuth: []

paths:
  /api/{user_id}/tasks:
    get:
      summary: List all tasks for authenticated user
      description: Returns a list of all tasks belonging to the authenticated user, ordered by creation date (newest first)
      operationId: listTasks
      tags:
        - Tasks
      parameters:
        - name: user_id
          in: path
          required: true
          description: User ID (must match authenticated user's ID from JWT)
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved task list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListResponse'
              example:
                tasks:
                  - id: "550e8400-e29b-41d4-a716-446655440000"
                    user_id: "7c9e6679-7425-40de-944b-e07fc1f90ae7"
                    title: "Buy groceries"
                    description: "Milk, eggs, bread"
                    completed: false
                    created_at: "2026-02-08T10:30:00Z"
                    updated_at: "2026-02-08T10:30:00Z"
                count: 1
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create a new task
      description: Creates a new task for the authenticated user
      operationId: createTask
      tags:
        - Tasks
      parameters:
        - name: user_id
          in: path
          required: true
          description: User ID (must match authenticated user's ID from JWT)
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
            example:
              title: "Buy groceries"
              description: "Milk, eggs, bread, and vegetables"
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
              example:
                id: "550e8400-e29b-41d4-a716-446655440000"
                user_id: "7c9e6679-7425-40de-944b-e07fc1f90ae7"
                title: "Buy groceries"
                description: "Milk, eggs, bread, and vegetables"
                completed: false
                created_at: "2026-02-08T10:30:00Z"
                updated_at: "2026-02-08T10:30:00Z"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/{user_id}/tasks/{task_id}:
    get:
      summary: Get a specific task
      description: Returns details of a specific task if it belongs to the authenticated user
      operationId: getTask
      tags:
        - Tasks
      parameters:
        - name: user_id
          in: path
          required: true
          description: User ID (must match authenticated user's ID from JWT)
          schema:
            type: string
            format: uuid
        - name: task_id
          in: path
          required: true
          description: Task ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      summary: Update a task
      description: Updates an existing task if it belongs to the authenticated user
      operationId: updateTask
      tags:
        - Tasks
      parameters:
        - name: user_id
          in: path
          required: true
          description: User ID (must match authenticated user's ID from JWT)
          schema:
            type: string
            format: uuid
        - name: task_id
          in: path
          required: true
          description: Task ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
            example:
              title: "Buy groceries and cook dinner"
              description: "Updated task description"
              completed: false
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      summary: Delete a task
      description: Permanently deletes a task if it belongs to the authenticated user
      operationId: deleteTask
      tags:
        - Tasks
      parameters:
        - name: user_id
          in: path
          required: true
          description: User ID (must match authenticated user's ID from JWT)
          schema:
            type: string
            format: uuid
        - name: task_id
          in: path
          required: true
          description: Task ID
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Task deleted successfully (no content)
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/{user_id}/tasks/{task_id}/complete:
    patch:
      summary: Toggle task completion status
      description: Updates the completion status of a task if it belongs to the authenticated user
      operationId: toggleTaskCompletion
      tags:
        - Tasks
      parameters:
        - name: user_id
          in: path
          required: true
          description: User ID (must match authenticated user's ID from JWT)
          schema:
            type: string
            format: uuid
        - name: task_id
          in: path
          required: true
          description: Task ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskComplete'
            example:
              completed: true
      responses:
        '200':
          description: Task completion status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /health:
    get:
      summary: Health check endpoint
      description: Returns API health status (no authentication required)
      operationId: healthCheck
      tags:
        - System
      security: []
      responses:
        '200':
          description: API is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "healthy"
                  timestamp:
                    type: string
                    format: date-time

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token issued by Better Auth. Include in Authorization header as: `Bearer <token>`

        Token must contain:
        - user_id: UUID of the authenticated user
        - email: Email address of the authenticated user
        - exp: Expiration timestamp (7 days from issuance)

  schemas:
    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Task title (required, 1-200 characters)
          example: "Buy groceries"
        description:
          type: string
          maxLength: 2000
          nullable: true
          description: Task description (optional, max 2000 characters)
          example: "Milk, eggs, bread, and vegetables"

    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Updated task title (optional, 1-200 characters)
          example: "Buy groceries and cook dinner"
        description:
          type: string
          maxLength: 2000
          nullable: true
          description: Updated task description (optional, max 2000 characters)
          example: "Updated description"
        completed:
          type: boolean
          description: Updated completion status (optional)
          example: false

    TaskComplete:
      type: object
      required:
        - completed
      properties:
        completed:
          type: boolean
          description: New completion status (required)
          example: true

    TaskResponse:
      type: object
      required:
        - id
        - user_id
        - title
        - completed
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: Unique task identifier
          example: "550e8400-e29b-41d4-a716-446655440000"
        user_id:
          type: string
          format: uuid
          description: Owner user identifier
          example: "7c9e6679-7425-40de-944b-e07fc1f90ae7"
        title:
          type: string
          description: Task title
          example: "Buy groceries"
        description:
          type: string
          nullable: true
          description: Task description (null if not provided)
          example: "Milk, eggs, bread, and vegetables"
        completed:
          type: boolean
          description: Completion status
          example: false
        created_at:
          type: string
          format: date-time
          description: Creation timestamp (ISO 8601)
          example: "2026-02-08T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp (ISO 8601)
          example: "2026-02-08T10:30:00Z"

    TaskListResponse:
      type: object
      required:
        - tasks
        - count
      properties:
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/TaskResponse'
          description: List of tasks
        count:
          type: integer
          description: Total number of tasks
          example: 1

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Error code
              example: "UNAUTHORIZED"
            message:
              type: string
              description: Human-readable error message
              example: "Invalid or expired JWT token"
            details:
              type: object
              description: Additional error details (only in development mode)
              additionalProperties: true

  responses:
    UnauthorizedError:
      description: Authentication failed - missing, invalid, or expired JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "UNAUTHORIZED"
              message: "Invalid or expired JWT token"

    NotFoundError:
      description: Resource not found or user does not have access
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "NOT_FOUND"
              message: "Task not found"

    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "VALIDATION_ERROR"
              message: "Invalid request data"
              details:
                field: "title"
                issue: "Title is required and cannot be empty"

    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "INTERNAL_SERVER_ERROR"
              message: "An unexpected error occurred"

tags:
  - name: Tasks
    description: Task management operations
  - name: System
    description: System health and status endpoints